% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_colourpicker.R
\docType{package}
\name{test_colourpicker-package}
\alias{test_colourpicker}
\alias{test_colourpicker-package}
\title{test_colourpicker}
\description{
This is a shiny project
}
\author{
\strong{Maintainer}: Emanuel Huber \email{emanuel.huber@pm.me}

}
