% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_colourpicker.R
\name{test_colourpicker}
\alias{test_colourpicker}
\title{Lauch the test_colourpicker  app}
\usage{
test_colourpicker(...)
}
\arguments{
\item{...}{Arguments to be passed to the function}
}
\value{
shiny application object
}
\description{
Lauch the test_colourpicker  app!
}
